efi: loader: fix NULL pointer dereference when deleting root volume handle
efi_file_delete() unconditionally dereferences fh->parent->fd, but the
root volume handle created by efi_open_volume() has fh->parent = NULL.
If a UEFI application calls Delete() on the root handle, this crashes.

Check for a NULL parent and return EFI_WARN_DELETE_FAILURE, as the root
volume cannot be deleted.

Link: https://lore.barebox.org/20260216084758.3548990-28-a.fatoum@pengutronix.de

+++ efi/loader/protocols/file.c
@@ -285,6 +285,11 @@ static efi_status_t EFIAPI efi_file_delete(struct efi_file_handle *file)
 
 	file_close(fh);
 
+	if (!fh->parent) {
+		free(fh);
+		return EFI_EXIT(EFI_WARN_DELETE_FAILURE);
+	}
+
 	ret = unlinkat(fh->parent->fd, fh->path, flags);
 
 	free(fh);
